Update int to long?

+++ Source/Podio .NET/Utils/ApplicationFields/CalculationApplicationField.cs
@@ -139,7 +139,7 @@ namespace PodioAPI.Utils.ApplicationFields
         ///     The app reference field, only when type is "reference" , "outgoing_reference"
         /// </summary>
         [JsonProperty("reference_field_id", NullValueHandling = NullValueHandling.Ignore)]
-        public int? ReferenceField_id { get; set; }
+        public long? ReferenceField_id { get; set; }
 
         /// <summary>
         ///     The id of the value field on the related app
